# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GenerateAiAssetRequestType = typeof GenerateAiAssetRequestType[keyof typeof GenerateAiAssetRequestType];


export const GenerateAiAssetRequestType = {
  image: 'image',
  story: 'story',
  video: 'video',
} as const;

export interface GenerateAiAssetRequest {
  type: GenerateAiAssetRequestType;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  prompt: string;
  /** @maxLength 120 */
  action?: string;
}

export type GeneratedAiAssetType = typeof GeneratedAiAssetType[keyof typeof GeneratedAiAssetType];


export const GeneratedAiAssetType = {
  image: 'image',
  story: 'story',
  video: 'video',
} as const;

export type GeneratedAiAssetStatus = typeof GeneratedAiAssetStatus[keyof typeof GeneratedAiAssetStatus];


export const GeneratedAiAssetStatus = {
  complete: 'complete',
  queued: 'queued',
} as const;

export interface GeneratedAiAsset {
  id: string;
  type: GeneratedAiAssetType;
  title: string;
  prompt: string;
  uri?: string;
  text?: string;
  mimeType?: string;
  status: GeneratedAiAssetStatus;
  provider: string;
}

/**
 * User-owned content saved only when cloud sync is enabled
 */
export interface SyncSnapshot { [key: string]: unknown }

export interface SyncState {
  enabled: boolean;
  /** @nullable */
  updatedAt: string | null;
  state: SyncSnapshot;
}

export interface SyncStateUpdate {
  enabled: boolean;
  state?: SyncSnapshot;
  clearCloudCopy?: boolean;
}

export type SocialConnectionProvider = typeof SocialConnectionProvider[keyof typeof SocialConnectionProvider];


export const SocialConnectionProvider = {
  facebook: 'facebook',
  messenger: 'messenger',
  telegram: 'telegram',
  instagram: 'instagram',
} as const;

export interface SocialConnection {
  provider: SocialConnectionProvider;
  displayName: string;
  description: string;
  readCapabilities: string[];
  sendCapabilities: string[];
  scopes: string[];
  connected: boolean;
  /** @nullable */
  authorizedAt: string | null;
}

export interface SocialConnectionConsent {
  enabled: boolean;
}