# lib/api-zod/src/generated/types/socialConnection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialConnectionProvider } from './socialConnectionProvider';

export interface SocialConnection {
  provider: SocialConnectionProvider;
  displayName: string;
  description: string;
  readCapabilities: string[];
  sendCapabilities: string[];
  scopes: string[];
  connected: boolean;
  /** @nullable */
  authorizedAt: string | null;
}