# lib/api-zod/src/generated/types/syncSnapshot.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * User-owned content saved only when cloud sync is enabled
 */
export interface SyncSnapshot { [key: string]: unknown }