# lib/api-zod/src/generated/types/syncState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncSnapshot } from './syncSnapshot';

export interface SyncState {
  enabled: boolean;
  /** @nullable */
  updatedAt: string | null;
  state: SyncSnapshot;
}