# lib/api-zod/src/generated/types/syncStateUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncSnapshot } from './syncSnapshot';

export interface SyncStateUpdate {
  enabled: boolean;
  state?: SyncSnapshot;
  clearCloudCopy?: boolean;
}